#23 - 스프링 데이터 JPA
스프링 데이터 JPA 에 대해서 학습해 보았다.
`findByName` 이라고 선언만 해주고 구현해주지 않았는데도 알아서 Name 을 찾는 쿼리문을 내준다. 정말 마법같았고 너무 신기했다.
