+++ src/main/java/com/example/jpaweb/repository/MemberRepository.java
@@ -1,33 +1,12 @@
 package com.example.jpaweb.repository;
 
 import com.example.jpaweb.domain.Member;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-import javax.persistence.EntityManager;
 import java.util.List;
 
-@Repository
-@RequiredArgsConstructor
-public class MemberRepository {
-    private final EntityManager em;
-
-    public void save(Member member) {
-        em.persist(member);
-    }
-
-    public Member findOne(Long id) {
-        return em.find(Member.class, id);
-    }
-
-    public List<Member> findAll() {
-        return em.createQuery("select m from Member m", Member.class).getResultList();
-    }
-
-    public List<Member> findByName(String name) {
-        return em.createQuery("select m from Member m where m.name = :name", Member.class)
-                .setParameter("name", name)
-                .getResultList();
-    }
-
+public interface MemberRepository extends JpaRepository<Member, Long> {
+    // select m from Member m where m.name = ?
+    // 이게 그냥 된다 Name 을 보고 그냥 되는 것이다.
+    List<Member> findByName(String name);
 }
